LPD-56533 Simplify
